# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: M&S Delivery System API
  version: 1.0.0
  description: Official REST API specification for M&S Delivery System (FastAPI backend on Vercel).
servers:
  - url: /api/v1
    description: Current API v1 base path
paths:
...
  /stores:
    get:
      summary: List stores
      operationId: listStores
      parameters:
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Store'
    post:
      summary: Create a new store (Admin only)
      operationId: createStore
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreCreateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
  /stores/{store_id}:
    get:
      summary: Get store details including menu categories and items
      operationId: getStoreDetail
      parameters:
        - name: store_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreDetail'
    put:
      summary: Update store details (Admin only)
      operationId: updateStore
      security:
        - BearerAuth: []
      parameters:
        - name: store_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreUpdateRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
...
  /uploads/signed-url:
    post:
      summary: Request short-lived signed upload URL for menu/store image
      operationId: getSignedUploadUrl
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignedUploadUrlRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedUploadUrlResponse'
  /uploads/finalize:
    post:
      summary: Finalize uploaded object
      operationId: finalizeUpload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeUploadRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    Store:
      type: object
      required: [id, name, address, latitude, longitude, is_active]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        image_url:
          type: string
        is_active:
          type: boolean
    StoreCreateRequest:
      type: object
      required: [name, address, latitude, longitude]
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        image_url:
          type: string
        is_active:
          type: boolean
          default: true
    StoreUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        image_url:
          type: string
        is_active:
          type: boolean
    MenuCategory:
      type: object
      required: [id, name, display_order]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        display_order:
          type: integer
    MenuItem:
      type: object
      required: [id, store_id, name, price, is_available]
      properties:
        id:
          type: string
          format: uuid
        store_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image_url:
          type: string
        is_available:
          type: boolean
...
    StoreDetail:
      type: object
      required: [store, categories, items]
      properties:
        store:
          $ref: '#/components/schemas/Store'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/MenuCategory'
        items:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
...
    SignedUploadUrlRequest:
      type: object
      required: [filename, content_type, size_bytes]
      properties:
        filename:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
    SignedUploadUrlResponse:
      type: object
      required: [upload_url, object_path, public_url, expires_in]
      properties:
        upload_url:
          type: string
        object_path:
          type: string
        public_url:
          type: string
        expires_in:
          type: integer
    FinalizeUploadRequest:
      type: object
      required: [object_path]
      properties:
        object_path:
          type: string
...